Further decomposes some GUI components
